Cambio en tabla Users, coneccion a la bd, endpoint get, getById, getByUsername y post

+++ database.sql
@@ -1,5 +1,5 @@
 
 CREATE TABLE Users (
-	Id numeric(38,0) IDENTITY(0,1) NOT NULL,
+	Id int IDENTITY(0,1) NOT NULL,
 	Username varchar(100) COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL
 );
